mvp/requirements.txt
Django==1.8.4
Jinja2==2.8
MarkupSafe==0.23
Werkzeug==0.10.4
appnope==0.1.0
bcrypt==2.0.0
cffi==1.2.1
decorator==4.0.2
django-appconf==1.0.1
django-compressor==1.5
django-debug-toolbar==1.3.2
django-extensions==1.5.7
django-jinja==1.4.1
django-jinja-bootstrap-form==4.0.3
django-longerusernameandemail==0.5.7
django-pgjson==0.3.1
gnureadline==6.3.3
gunicorn==19.3.0
ipython==4.0.0
ipython-genutils==0.1.0
path.py==8.1.1
pexpect==3.3
pickleshare==0.5
psycopg2==2.6.1
py==1.4.30
pycparser==2.14
pytest==2.8.0
pytest-django==2.8.0
simplegeneric==0.8.1
six==1.9.0
sqlparse==0.1.16
traitlets==4.0.0
wsgiref==0.1.2
